Switching to "template" plug-in type.
Using the "stylesheet" brunch plug-in type, does not emit junk
into compiled javascript, however, it does cause down-stream
`onCompile()` hooks like `auto-reload-brunch` to think that the CSS
has changed, and not the page itself. Which will require a manual refresh.

It appears that the "template" type has been fixed so that it no longer
emits the wrapper, if the returned result is empty, and also triggers
the proper down stream changes for `onCompile()`.